+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file contains the list of changes between different versions of Upload Exte
 * Fix update tooltip for mobile devices.
 * Fix updating extensions with removed `ext.php`.
 * Fix code selection functionality in file trees.
+* Fix version check functionality.
 
 # Version 3.1.2
 ## Changes since 3.1.2-RC2